♻️ refactor: refactor trace type (#8699)
* refactor trace type

* refactor lobechat type

* refactor types
