+++ packages/types/package.json
@@ -2,19 +2,11 @@
   "name": "@lobechat/types",
   "version": "1.0.0",
   "private": true,
-  "exports": {
-    "./*": {
-      "types": [
-        "./src/*.ts",
-        "./src/*/index.ts"
-      ],
-      "default": [
-        "./src/*.ts",
-        "./src/*/index.ts"
-      ]
-    }
-  },
+  "main": "./src/index.ts",
   "dependencies": {
-    "@lobehub/chat-plugin-sdk": "^1.32.4"
+    "@lobechat/web-crawler": "workspace:*",
+    "@lobehub/chat-plugin-sdk": "^1.32.4",
+    "type-fest": "^4.41.0",
+    "zod": "^3.25.76"
   }
 }
