--- packages/types/src/share.ts
@@ -1,7 +1,0 @@
-export interface ShareGPTConversation {
-  avatarUrl?: string | null;
-  items: Array<{
-    from: 'human' | 'gpt';
-    value: any;
-  }>;
-}
