+++ packages/types/src/trace/index.ts
@@ -1,1 +1,2 @@
 export * from './action';
+export * from './enum';
