+++ src/app/(backend)/webapi/models/[provider]/pull/route.ts
@@ -1,4 +1,4 @@
-import { ChatErrorType } from '@lobechat/types/fetch';
+import { ChatErrorType } from '@lobechat/types';
 
 import { checkAuth } from '@/app/(backend)/middleware/auth';
 import { ChatCompletionErrorPayload, PullModelParams } from '@/libs/model-runtime';
