+++ src/app/(backend)/webapi/trace/route.ts
@@ -1,6 +1,6 @@
+import { TraceEventType } from '@lobechat/types';
 import { after } from 'next/server';
 
-import { TraceEventType } from '@/const/trace';
 import { TraceClient } from '@/libs/traces';
 import { TraceEventBasePayload, TraceEventPayloads } from '@/types/trace';
