+++ src/types/locale.ts
