+++ src/utils/fetch/__tests__/parseError.test.ts
@@ -1,4 +1,4 @@
-import { ErrorResponse } from '@lobechat/types/fetch';
+import { ErrorResponse } from '@lobechat/types';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import { getMessageError } from '../parseError';
